Fairing baseplate masses updated
Feel free to revert this if it's out of line, but I noticed some of the
PLF bases seemed extraordinarily lightweight. This is also a physics
problem: a too-light part in between two heavy parts (say, the main
tank, and the payload) will lead to wobbliness and breaking rockets.
Certainly at the least the mass of a baseplate should go up with the
square of its radius, as that's assuming it's entirely hollow!