Create RO_DMagic.cfg
RO compatible science parts for rovers and orbital observations. I have made some of them a bit heavier and removed the option to scale them, because it didn't work properly and I think that it really isn't needed. Only the Solar Collector and Soil Moisture Reader can be made bigger with as a "stack" configuration.