Fix Mercury retros and posigrade motors
The given Isp is wrong, given known thrust, burn duration, and fuel mass (and delta V). Isp has been calculated to be 69.1s in vacuum. Masses for the entire retro pack (decoupler, pack, and the three posigrade motors) is correct at 237kg, and thrust for the posigrade motors was corrected to 400lbf.