FASA Mercury Changes
Changed the behaviour of a number of Mercury capsule parts. The capsule
cap will now pop off on re-entry. The RCS pods are now attached to the
base capsule corrected, RCSFX functions. LES now stages as the Apollo
LES does.

Added a txt file for keeping track of issues ID'd and corrected as I go
through the capsules. Will be removed on completion of my pass through.