Apollo CSM LEM Changes
Rebalanced the SM RCS around the CSM centre of mass to improve balance.
Works best when CM RCS is toggled off. (weld)

Added CM RCS models to the CM in the appropriate spots. (weld)

Added LEM docking cone and docking function to LEM ascent stage. (weld)

Moved the attachment point in LMA base to support a separator at the
base of the LEM. To be used after transposition and docking to seperate
LEM from S-IVb.