Increase LH2 in Service Module Tanks
Current amount of LH2 in the Service Module only lasts 8 days, 16 hours; not enough to provide electricity for the duration of most missions to the moon. This update adds enough to last 14 days, putting it in line with the actual duration of missions, as well as the amount of food on the Service Module (which also lasts 14 days).