Tweaks
I made the engines a little more efficient (matches what Astronautix
seems to say) and now the historical flight profile seems to work.

Still needs a few fixes:

LS needs balancing, I messed up and did calculation using Kerbin days.

The service module (responsible for TEI) needs to be changed to stage as
an engine, I forgot how to change that. It also lights on the launch pad
or mid-flight.