Replace RSRMV with SRMU, switch model for SRMU and UA1207
At some point the KW SRMU was replaced with the RSRMV, though the model size was never changed. Since the RSRMV is part of the Squad pack, switching it back to SRMU makes sense. Also switched models for the SRMU and UA1207 so the 7-segment UA1207 gets the model with more segments and the 3-segment SRMU gets the model with fewer.