adding sxt-rd107/108
Clone the existing RD-170 model, rescale. Intended for use with RP-0 to allow the construction of R-7 rockets.

One upgraded config placed, price and tech-tree spot estimated. This may require refinement in the future.

Engines gimbal so that the craft is controllable without verniers, as MechJeb control of actual verniers is a bit poor.


There will also be an RP-0 PR for initial placement in the techtree for RP-0.