Update RO_FASA_ApolloCSM.cfg
For the CM change, as it stood the CO2 Scrubber only scrubbed about 95% of the CO2 out of the air.  That meant you were constantly building up CO2 with no way to actually remove it.  And if the scrubber every stopped working for a short time, you'd never be able to recover from that.  We know from historical events (Apollo 13) that the scrubbers were actually able to reduce the amount of CO2 in the air.  The new value will convert 105% of CO2 while generating the same amount of O2.

For the SM change, the old fuel cell did not produce enough electricity to power the CSM plus the high gain antennae.  The new values should be enough for both.