Update RO_FASA_ApolloCSM.cfg
We know from historical records that the Apollo CSM had a camera system.  My guess is they had a lot of other scientific equipment as well but most of that can be added into the equipment bays.  We don't have a camera (or material lab in stock) that will properly fit into the equipment bay so I think adding the experiment to the SM directly is a valid way to correct this issue.