Update RO_FASA_Atlas.cfg
Added a configuration for the LA89-NA-5 which is part of the Atlas MA-2 booster.  While the booster is only slightly more powerful than the NA-3 version (the MA-1 booster), it is the correct values for recreating the Atlas D rocket.
I've also relabeled the LR105-NA-6 to (hopefully) show that it was used in both the MA-2 and MA-3 engines.  There is no game difference between the LR105-NA-5 and LR105-NA-6 (the NA6 had a slightly higher chamber pressure) so no reason to actually include a separate configuration.