Update RO_FASA_Saturn.cfg
Historically, once the Apollo CSM and LEM have left the Saturn IVB stage behind (at least on the mid and late Apollo flights), Houston was able to remotely control the stage to alter it's trajectory so that it could impact the lunar surface.  This change adds remote control capability and antenna (the same Omni that the CM has) to the Saturn Instrument Unit to allow for this capability.