Update RO_FASA_Atlas.cfg
We get the -3 at Basic Orbital Rocketry (unchanged), the -5 at Mature Orbital Rocketry (where we used to get the -6), the -6 at Heavy orbital Rocketry (new), the -7 at Refined Rocketry (unchanged) and the RS at "experimentalRocketry" (unchanged).  Presumably there is something done in RP-0 to move the RS to Advanced Rocketry (veryHeavyRocketry) but I don't think that exists on the tree in RO.  Fortunately, since the RS does not need to be moved, it should be fine as is.