Increase Gemini chute cover maxSkinTemp
When in re-entry mode, the Gemini capsule bobs around a bit (and also changes pitch as the atmosphere thickens). This causes the nose cone to be intermittently exposed to re-entry heat, and often explode, leading to the whole craft to be lost.

By increasing the maxSkinTemp of this part, it won't explode. Tested during controlled LEO and lunar-return velocity returns.