Update RO_FASA_Titan.cfg
Adjusted the dimensions and weight on the Centaur D/D1, D2 and D3 tanks.  I've used values as presented in astronautix  for the sake of consistency.  I did compare values to what is presented in b14643.de and my results are comparable.  But I did notice that b14643.de uses the same basic statistics for similar, but not identical, stages.  For instance, there are five different versions of the Atlas II presented, each with their own slightly different Centaur stage (each uses a different RL10A engine set) but the life off weights and propellant weights never change.  Seems unlikely that the RL10A-3-3A had the exact same weight as the RL10A-4-1N.  In the end, I tried to be consistent but also went with lower weight values on the assumption that minor changes could be dealt with by including a small amount of leadbalast.
The final tank height includes the tank and engines.
The final tank weight includes the tank, engines, payload fairing base and 0.03t worth of RCS thrusters.

I've also added a configuration for the Centaur D5 (used with the Atlas V) and the Centaur T (used with the Titan IV).  Information is based on the already existing Centaur tanks with weights and dimensions coming from astronautix and b14643.de

Finally, I've included a RCS thrusters for centaur roll control.  From various wiki sites, it seems that a standard Centaur setup included 8 thrusters for roll and attitude control and 4 slightly more powerful thrusters for ullage.  The ullage thrusters can easily be recreated using existing Attitude Jets but I've always run into problems properly positioning single jets for roll/attitude control.  To resolve this, I've used the FASA_Gemini_RCS_Thrusters as a model for small, Hydrazine powered attitude thrusters for the Centaur.  I've reduced the weight of the original (0.015 down to 0.006) since the centaur versions are weaker but I have no documentation for how much they should actually weigh.