New configs
NOTE: There is the need to convert tabs --> spaces. Reviewing of all the
changes must be done as a first step because the conversion will make
GitHub think that the configs are brand new.

Conversion will be done after the OK for the current state of the
configs is given.