AJ10-118F/K Delta configurations
Add engine configurations for the AJ10-118F and AJ10-118K.  These engines were used on the "Delta F" and "Delta K" second stages for Delta I and Delta II rockets.  I've added these configurations to the existing SXTAJ10Mid engine because they were advances on the older AJ10-118E.  That may be accurate for the 118F since the Delta F second stage was similar in size to the older Delta E second stage.  But it's possible the 118K should be a completely separate engine since the Delta K was larger.  Trouble is I can't find documentation that shows the dimensions of the engine itself.