changes to radiator
based on in-game experience and discussion with NathanKell

@NathanKell, please point out of any of my comments are incorrect. I figured it would be useful to have some documentation of these variables.

Also, according to NathanKell (if I interpreted this correctly) radiatorMax sets the priority of a part to be cooled, higher is better. above 0.01, below 0.99. I set some tanks to 0.9 (2 tanks on one craft) and the habitation module to 0.35 and it didn't seem to behave as expected, so I didn't commit that to this PR