Reorganization of Engines, Plumes fixes
Changes in SSTU affected the RO clusters substantially. After I worked
out the tricks required to get them to work, I standardized the process
and applied it to a few families of engines.

I gave those engines their own files, as there are so many engines and
clusters in SSTU now, that is the organizational scheme that probably
makes the most sense.

SSTU changes also broke the RO RealPlumes for engine cluster.  This PR
includes fixes for those as well (for the engines I have previously made
RealPlumes for)