Update RO_Squad_Engines.cfg
Update to correct LR105 mass to account for no fuel tank on the Squad version.

Update the LR79 engine data to closer reflect data from the NASA handbook for this engine.  The numbers I'm using for the LR79-9 and LR79-11 are pulled directly from the 1961 handbook.  A1Ch1 suggested this for the LR79-9 but suggested a -4% deviation for the LR79-11.  As I'm unclear why the deviation was added to the 2nd configuration and not the first, I elected to drop the deviation completely and just use the source data as it was provided.