CO2 Scrubber for Skylab
Adding a TACLS CO2 Scrubber to the Skylab since I have to assume it had one.  I've used the same scrubber that we have on the FASAApollo_CM part, though I've adjusted the conversaionRate up to 6 to account for the 6 crew capacity on the Skylab.  This will result in twice as much CO2 and LithiumHydroxide being used, with twice as much Waste being generated.

I've also reset the LS quantities so there should be 300 days worth of LS (1 day worth of CO2), including running the scrubber.