More Big G Changes
I'm changing the texture used for the Big G Retro Module.  The original texture didn't handled the 3.9m to 4.6m taper as well as the texture I'm using now.  I also read (http://www.astronautix.com/craft/bigemini.htm) that the retro module only included the retro engines, plus water and oxygen.  It didn't include any attitude control thrusters like the comparable module on the standard Gemini setup.  I've therefore removed the RCS module from this part and I've added a 2nd "fuel tank" setup specifically for life support to store the bulk of the flights water supply and backup O2.  This module includes textures for four "Gemini Mini SRB" to act as the nozels for the retro engines.  About the only real issue with this texture (that I can see) is that it is no longer "hollow".

I'm changing the texture used on the Big G Service Module so that we now only use the FASAGeminiBigGDock graphic.  I also added four Apollo RCS thruster textures since the same page I cited above says the service module includes the attitude control system.  The illustrations I've found seem to agree.  I've also removed the bulk of the water from this module (moved to the retro module) and reduced the storage volume accordingly.  The ModuleRCS thrust on this part has been doubled as compared to the RCS on a standard Gemini.  The tests I've run so far seem to work well with this thrust setting.

Lastly, I've altered the size of the Big Gemini Docking Ring to more closely match the dimensions I'm seeing on various illustrations (see above website).