SLA/IU Collider workarounds
Based on numerous tests I've run, I'm 99.9% sure that the collider problem is on the IU texture itself.  I tried creating a single part that was made up of the two "FairingWall3m" of the standards SLA and the "Apollo_IU".  But when the LEM separates from the SLA, it is still picking up phantom momentum as it's "pushed" out of a texture. We definitely someone that can look at the collider on the texture itself and fix it.  But in the mean time I've got an update to the standard SLA so you no longer need a separate decoupler for the LEM.  Those changes include:
1) I found some documentation that puts the SLA at "7-foot long" which works out to 2.1336m.  The existing part is 2.07m.  This doesn't seem like much, but every mm reduces the amount of phantom momentum when the LEM separates.  So I've corrected the height of the basic SLA to be 2.1336m long.
2) I've slightly adjusted the fairing, top and bottom attachment points to account for the tiny increase in size and also to remove any clipping the fairing may have with the CSM.
3) I've raised the LEM attachment point to account for no longer needing an extra decoupler.

I've also added two new interim parts.  The first is a 2.6m long SLA.  It's not drastically larger than the standard part (about +22%) but does cut the phantom momentum by about half.  And there is no phantom momentum if you build your own LM so nothing is below the descent engine.
The second is a 3m long SLA.  It's large enough that the normal LM doesn't hang below it resulting in no phantom momentum upon LEM separation.