Skylab storage capacity
In a previous PR we corrected the amount of life support that is included with Skylab to reflect what the actual Skylab included, but we didn't adjust the available storage volume.  The 85.5kL of space meant you could easily setup the Skylab with nearly 2 years worth of life support.  I'm not sure if that volume was made up or if there was some documentation to go along with it.  Barring any documentation, I've lowered the volume so that there's only about 45% additional storage capacity.

I've also altered the initial RCS fuel capacity.  According to http://pages.erau.edu/~ericksol/projects/issa/skylab.html#Skylab Systems the Skylab's "Thruster Attitude Control Subsystem" was powered by "Cold gas propellant (nitrogen)" but the Skylab we have uses MMH/NTO for it's RCS thrusters.  I'm not sure if that was done intentionally or not.  And I didn't want to mess with the existing RCS without knowing exactly why MMT/NTO was used in place of the actual Nitrogen system.  But I did want to make sure that the same quantity of fuel was available so I've adjusted the amount of MMH and NTO available to approximately match the weight of Nitrogen that would have been included.  Here's my math:

http://pages.erau.edu/~ericksol/projects/issa/skylab.html#Skylab Systems says Skylab included "22 propellant storage spheres" and that each sphere "stored [nitrogen] at 3,100psi at 100 degrees F in 4.5 cubic foot titanium spheres ".
According to http://siei.org/usefulequations.html, "A K-cylinder holds about 1 cubic foot [at STP, and]....A K-cylinder of nitrogen (at 3000 psi) holds roughly 250 scf"
Based on the above, the Skylab has 22 spheres, each holding approximately 1125 standard cubic feet (scf) of nitrogen.  That works out to 24750 scf, or 700841.95L of nitrogen.  That is roughly 876.75kg of nitrogen based on the density provided in CommonResources.cfg.
At a 45.2%/54.8% MMH/NTO ratio, 332.35MMH + 402.95NTO works out to 876.75kg.