Skylab notations and modification
Added notation to the NTO/MMH values to explain where the amounts came from.

Adjusted the ModuleDataTransmitter values for the Skylab antenna so they match the SM antenna.  I'm not positive how powerful the actual transmitter would have been but I'm guessing it would have been similar to what was available on the Apollo.