BigG adjustments
I've increased the amount of fuel in the FASABigGeminiRetroModule part.  This gives this part the same delta-V when attached to the Big G that the FASAGeminiUtilitySasRcs gives the standard Gemini.

Because of the animation on the Big G docking port, I needed to increase the acquireRange and captureRange for docking.  These values have to include the length of the extended docking node.

Finally, I've tweeked the position and angle of the RCS thrusters on the FASAGeminiBigGDock to minimize unwanted torque/thrust during attitude control and translation maneuvers.  The new position/angle assumes that the guidance part's RCS is turned off (since it's only supposed to be used during re-entry).  The new position/angle also assumes just stock "fuel".  If you load up the FASAGeminiBigGDock with 39k liters of extra "fuel" (whether that be extra life support, extra fuel, extra lead, extra whatever) you will likely need to add additional small RCS thrusters to account for the extra mass of the BigG.