Rescaled the Mk1-2 pod and Mk2 Lander Can IVA's.  Should work for both stock (RO) and Ven's along with RPM varients.
Word of caution:  Other ships using these IVA's that aren't scaled the same way as the
aforementioned ships might not line-up properly in the IVA.  But I figure it cannot be
any worse than *not* having it scaled this way.  For example, if this IVA was scaled to
1.6, but your cockpit is scale to 1.72222 then my rescale will arguably be better than
the unscaled IVA, but slightly off.