Created preliminary config for French Vexin engine family
WIP!

I cross-checked all parameters except for engine mass, which was nowhere to be found (Astronautix has 175 for the semi-related Vexin A used on Europa, but I don't trust it), 450 kg is a placeholder. For the ignition resource, sources mention "fantol", whatever that is. I just made it use an equivalent amount of Furfuryl alcohol for now.