Mercury heatshield 1.1 tweak
During a 'standard re-entry' from a 250km x -314km orbit, 52 of 125 ablator remain and a peak skin temp of 1950K is achieved.

In my test, using these specs for the mk1 capsule results in a peak skin temp of 1780K is achieved and 116/200 ablator remains. Given that both of these capsules will start to burn at ~2150K, the mk1 is left with much more temperature overhead, so it has a much easier re-entry.

This PR should wait on confirmation testing that the values achieve the same or very similar results for someone else. At that point, please let me know and I will go adjust all the other values that should be changed as well.