SSTU - Adjust diameter increment
In SSTU, the major increment is set at 2x the minor increment, which
means that setting the minor increment to 0.1 means that the major
increment has to be adjusted by 0.2m, which makes it difficult to
actually change the diameter by any significant amount.  Testing
suggests that relatively fine adjustment is still possible.