RN fix RT maxq problem
-The current values were breaking antenna off during ascent on probes
that have antenna as part of the main mesh.

This value works for when the antenna is a separate part and hidden
under a fairing, but on my parts where the antenna is part of the main
mesh, if the main body of the probe, like skylab, is exposed to the air
during launch it breaks the antennae off even though they are under the
fairing.