RN rcs fixes
@Nathankell these are the fixes I have come up with for a few parts.

All of these parts needs RCS enabled by default because they fit a
certain criteria:
1) They have a combined solid stage/decoupler and rcs module
2) They are the first part on the rocket stack that needs rcs enabled,
so enabling it by default makes sense
3) They need RCS for ullage of a liquid stage and the rcs and liquid
engine are combined on one part

every one of these parts fits the 2nd criteria so having rcs enabled by
default is perfectly acceptable, but that alone was not enough for me.

I omitted a bunch of parts as I decided they need to fit at least 2 of
those options to be modified to be on by default.