Add Non-notional NTR engines to RO using stock/vens (#1720)
* Add non-notional Nuclear Thermal Rocket engines adding 10 NTR's from the '60s through '90s
These include only non-notional versions with the exception of the SNTP engine which was tested in part but not run as a whole assembly. Further work will update core NTR's in RO/RP-0 to closer to expected values. after that the Ven's engines will be given configs to bring them in line with real NTR's followed by configs for Kerbal Atomics.

* move file to better folders
move to proper locations

* PAP UPDATE: Removed all spaces in the config files and replaced them with hyphens '-' in order to work better with all aspects of TF and MM and KSP. Also fixed the names of some of the configs as they have to reference valid configs. THANKS BORNHOLIO!