Move BDB X-405 turbopump config into separate pass
The X-405 is supposed to have 2 engine modules - one for the "main"
engine, and one for the turbopump exhaust (see the SXT X-405). The
Bluedog one, though, was engine up with 3 engine modules - one for the
"main" engine, and two for the turbopump exhaust, because the config was
adding another turbopump exhaust config instead of editing the existing
one (the BDB version provided a turbopump engine module, while the SXT
one did not).

Also deleted keys that were already set to the specified values in the
original BDB config. Had to add :HAS[#thrustVectorTransformName[]] to
the section of the config editing the turbopump exhaust since the part
has two ModuleEngines*.