Update SSTU_Tanks.cfg
This should fix the problem with the Modular Standard Fuel Tank as well as update the Modular Upper Stage to work with 1.3.1. 

I noticed that there is a problem with the SSTU/ModIntegration/RealFuels/RF.cfg file.  It's still referencing the old way the MUS was named.  There is a PR on the SSTU github (#678) but until it's released, I needed to add a version of it here so that the MUS would get the correct fuel tank options.