SSTU Engine Mass Fix
Work around for the engine clustering mass bug. Sets the stock mass parameter to the appropriate value and ignores the origMass parameter. As a consequence, engines with multiple variants have a fixed mass. RD-107 and RD-108 engines are excluded, as I personally never cluster these engines and would prefer to preserve the mass switching with upgraded variants.