Fixes decay of RTGs over time
The decay of RTGs over time doesn't appear to be working properly, this fixes that and removes the Resource Converter module and replaces it with NFE's Radioisotope Generator module. 

Decay rate was taken from actual values taking both the decay of the fuel and the thermocouples. 

Note: This patch will only apply if both RO and NFE are installed.