Resize RestockPlus RCS blocks
They were maybe half as big as the "stock" RO blocks
of equal thrust/isp. Eyeballed the 28, 69 and 138 blocks
to be roughly the right size.

Didn't touch bigger sizes because I don't have them unlocked;
those using the 'mini' models were the most problematic anyway.

A bit of compromise involved: making the nozzle sizes match
tended to make the overall part bigger, so I erred on the side
of smaller