Fix KerbCan IVA scaling
Yet another case of comically oversized IVA. At this point I'm guessing
KSP just changed the way IVAs were scaled and values that used to
be fine suddenly weren't.

Only tested with the RPM variant on the 1-crew KerbCan; hopefully
the change is valid for all permutations. If not, something more
sophisticated will be needed