Add XLR43 Engine_Config
Add engine Config for XLR43,  navaho booster and predecessor of all american kerolox engines. It started as the XLR43-NA-1, which the NAA 75-110 was derived from. It later was upgraded with a bipropellant gas generator and burning Ethanol 90 as the XLR43-NA-3, which flew on Navaho tests and became the Altas boosters when upgraded to burn kerolox (as LR43)