CryoEngines update (#2278)
Edited the Vulcain engine to only use the Vulcain 2 config since that's what the part is modeled after.

Even though the RS-68 roll control exhaust animation is broken, it works just fine. Couldn't find what the deflection angle is, so used 15 degrees (copied from ROEngines).

Rescaled the RS-68 based on the boattail diameter (so it fits the Delta IV's CBC) instead of the nozzle exit diameter. As a result, the engine is slightly smaller than the real RS-68.

Since there's no LE-9 default config available, I used the LE-7 one for the Fuji, but kept the scale of the engine according to the LE-9 nozzle exit diameter (1.59m according to the NF_Realism team).