Make 1.25m Cockpit 1.25m
Regression from restock/vens cleanup: the use of a vensnewparts model
was removed, but its rescaling was kept, so it ended up enlarging a
1.25m part by another 1.25x

- disable the entire VSR block, it really wants to depend
  on Ven's New Parts. We'll stick to stock model at stock size for now
- bring back the model reference so it's ready to go when the VNP issue
  gets resolved
- don't delete the light animation, because why??

Tested with Ven's New Parts installed but not Ven's Stock Revamp.
NOT tested with both installed; would probably be a good idea for
someone to try it.