Basic support for CarnationRED Flexible Parts (#2424)
* Basic support for CarnationRED Flexible Parts

- extract proctanks "set availableTypes" into reuable helper
- configure the core CRFP part to expose the classic
  Default/Fuselage/etc realfuel tank types
- add additional parts corresponding to procparts Conventional,
  Integral, etc types

For cylinders, dry mass and available volume seem to match procparts
equivalents in all cases. For more exotic shapes, we get what we get.

One omission: a pure, tankless "structural" part. Not a priority,
and couldn't quite see how to use this to configure mass scaling:
https://github.com/CarnationRED/CarnationREDFlexibleParts/blob/c5bad4fa060b07ff6dc0cd3bcd126acd0ddede2a/Main/ModuleCarnationVariablePart.cs#L809-L810

* clear CRFP tech limits

* 2m half-circle SectionCorners

* Add ModuleToggleCrossfeed to CRFP tanks

For parity with ROTanks, and because why not