ROWaterfall updates (#2455)
* ROWaterfall: BREAKING: delete all handwritten modules without mark
Establish a convention for the treatment of parts that ship their own
Waterfall config: If an ROWaterfall config is being written for such
a part, it is assumed that the original should be deleted (unless the
`rowaterfallKeep` key is added to indicate otherwise).

* ROWaterfall: BREAKING: Rename specialty audio templates.
The (new) policy is that there should only be one version of specialty
effects so that their interpretation is uniform across RO parts. Thus,
there is no need to number specialty templates.
The templates were made before this policy was established, hence the
retroactive renaming.

* ROWaterfall: Template art updates.
* Use a less harsh startup sound for Gamma and Stentor audio templates.
* Reduce blur strength for SRM template.
* Redone vacuum transition for SSME template.

* ROWaterfall: Fix glow resize for some templates.

* ROWaterfall: Nuclear engine pass.
* Add new template that's more fanned out and easier to place.
* Fix squish bug on all configs.
* Add support for VSR.

* ROWaterfall: Add support for Astris, NSTAR, and RCS blocks from VSR.

* ROWaterfall: Fix squish bug on generic thrusters.

* ROWaterfall: More SSME template updates.
* Increase plume size in vacuum slightly.

* ROWaterfall: Update attributions.

* ROWaterfall: Make AutoConfig more robust.
Edit-or-create `MainPlumeVariant`s to avoid duplication in case there
are already manually written variants of the same name.

* ROWaterfall: Add configs for Coatl Aerospace.

* Waterfall: Add config for KDU414.

* Waterfall: fix plume positioning for Coatl's 0.5kN generic thruster

* ROWaterfall: BREAKING: set plume variant with `rowaterfallVariant`
This is in place of having configs set `b9psSubtypeName` directly,
because the behavior of that key might change in the future to a form
that would force implementation details of ROWaterfall to be exposed.